agent-driven: Move get_unicodeset_property to ucd_helper.rs
